fix: allow longer timeout for complex load scenarios
We think 2000 was arbitrarily chosen as, in a majority of cases, it shouldn't take longer than two seconds to load a font.

Storybook takes longer than two seconds, hence it always times out due to the hard-coded value changed here.

In the wild, proxies, or other things that can slow a network connection to a crawl. Our testing showed that 10 seconds was a long enough timeout for font loading to succeed.

In the long run it would probably be a good idea to parameterize this
timeout as well, for now this works.